+++ geopackage-ios/extension/link/GPKGFeatureTileLink.m
@@ -1,0 +1,19 @@
+//
+//  GPKGFeatureTileLink.m
+//  geopackage-ios
+//
+//  Created by Brian Osborn on 2/4/16.
+//  Copyright © 2016 NGA. All rights reserved.
+//
+
+#import "GPKGFeatureTileLink.h"
+
+NSString * const GPKG_FTL_TABLE_NAME = @"nga_feature_tile_link";
+NSString * const GPKG_FTL_COLUMN_PK1 = @"feature_table_name";
+NSString * const GPKG_FTL_COLUMN_PK2 = @"tile_table_name";
+NSString * const GPKG_FTL_COLUMN_FEATURE_TABLE_NAME = @"feature_table_name";
+NSString * const GPKG_FTL_COLUMN_TILE_TABLE_NAME = @"tile_table_name";
+
+@implementation GPKGFeatureTileLink
+
+@end