+++ geopackage-ios/tiles/overlay/GPKGFeatureOverlay.h
@@ -10,31 +10,17 @@
 #import "GPKGBoundingBox.h"
 #import "GPKGFeatureTiles.h"
 #import "GPKGProjection.h"
+#import "GPKGBoundedOverlay.h"
 
 /**
  *  Feature Tile Overlay which draws tiles from a feature table
  */
-@interface GPKGFeatureOverlay : MKTileOverlay
+@interface GPKGFeatureOverlay : GPKGBoundedOverlay
 
 /**
  *  Feature tiles
  */
 @property (nonatomic, strong) GPKGFeatureTiles *featureTiles;
-
-/**
- *  Min zoom
- */
-@property (nonatomic, strong) NSNumber * minZoom;
-
-/**
- *  Max zoom
- */
-@property (nonatomic, strong) NSNumber * maxZoom;
-
-/**
- *  Web mercator bounding box
- */
-@property (nonatomic, strong) GPKGBoundingBox * webMercatorBoundingBox;
 
 /**
  *  Initialize
@@ -44,13 +30,5 @@
  *  @return new feature overlay
  */
 -(instancetype) initWithFeatureTiles: (GPKGFeatureTiles *) featureTiles;
-
-/**
- *  Set the bounding box, provided as the indicated projection
- *
- *  @param boundingBox bounding box
- *  @param projection  projection of the bounding box
- */
--(void) setBoundingBox: (GPKGBoundingBox *) boundingBox withProjection: (GPKGProjection *) projection;
 
 @end